+++ weixin-java-open/src/main/java/me/chanjar/weixin/open/api/WxOpenMaService.java
@@ -236,18 +236,6 @@ public interface WxOpenMaService extends WxMaService {
    */
   String API_AUDIT_UPLOAD_MEDIA = "https://api.weixin.qq.com/wxa/uploadmedia";
 
-  /**
-   * 小程序管理-获取公众号关联的小程序
-   */
-  String API_WX_AMP_LINK_GET = "https://api.weixin.qq.com/cgi-bin/wxopen/wxamplinkget";
-  /**
-   * 小程序管理-关联小程序
-   */
-  String API_WX_AMP_LINK_CREATE = "https://api.weixin.qq.com/cgi-bin/wxopen/wxamplink";
-  /**
-   * 小程序管理-解除已关联的小程序
-   */
-  String API_WX_AMP_LINK_UN = "https://api.weixin.qq.com/cgi-bin/wxopen/wxampunlink";
 
   /**
    * 小程序管理-查询小程序版本信息
@@ -676,55 +664,6 @@ public interface WxOpenMaService extends WxMaService {
    * @return
    */
   WxMaAuditMediaUploadResult uploadMedia(File file) throws WxErrorException;
-
-  /**
-   * <pre>
-   * 获取公众号关联的小程序
-   * 请求方式：POST(HTTPS)
-   * 请求地址：
-   * <a href="https://api.weixin.qq.com/cgi-bin/wxopen/wxamplinkget?access_token=TOKEN">https://api.weixin.qq.com/cgi-bin/wxopen/wxamplinkget?access_token=TOKEN</a>
-   * 文档地址：
-   * <a href="https://developers.weixin.qq.com/doc/oplatform/Third-party_Platforms/2.0/api/Official__Accounts/Mini_Program_Management_Permission.html">https://developers.weixin.qq.com/doc/oplatform/Third-party_Platforms/2.0/api/Official__Accounts/Mini_Program_Management_Permission.html</a>
-   * <pre>
-   * @return 公众号关联的小程序
-   */
-  WxAmpLinkResult getWxAmpLink() throws WxErrorException;
-
-  /**
-   * <pre>
-   * 关联小程序
-   * 关联流程（需要公众号和小程序管理员双方确认）：
-   * 1、第三方平台调用接口发起关联
-   * 2、公众号管理员收到模板消息，同意关联小程序。
-   * 3、小程序管理员收到模板消息，同意关联公众号。
-   * 4、关联成功
-   * 等待管理员同意的中间状态可使用“获取公众号关联的小程序”接口进行查询。
-   * 请求方式：POST(HTTPS)
-   * 请求地址：
-   * <a href="https://api.weixin.qq.com/cgi-bin/wxopen/wxamplink?access_token=TOKEN">https://api.weixin.qq.com/cgi-bin/wxopen/wxamplink?access_token=TOKEN</a>
-   * 文档地址：
-   * <a href="https://developers.weixin.qq.com/doc/oplatform/Third-party_Platforms/2.0/api/Official__Accounts/Mini_Program_Management_Permission.html">https://developers.weixin.qq.com/doc/oplatform/Third-party_Platforms/2.0/api/Official__Accounts/Mini_Program_Management_Permission.html</a>
-   * <pre>
-   * @param appid 小程序 appid
-   * @param notifyUsers 是否发送模板消息通知公众号粉丝
-   * @param showProfile 是否展示公众号主页中
-   * @return 响应结果
-   */
-  WxOpenResult wxAmpLink(String appid, String notifyUsers, String showProfile) throws WxErrorException;
-
-  /**
-   * <pre>
-   * 解除已关联的小程序
-   * 请求方式：POST(HTTPS)
-   * 请求地址：
-   * <a href="https://api.weixin.qq.com/cgi-bin/wxopen/wxampunlink?access_token=TOKEN">https://api.weixin.qq.com/cgi-bin/wxopen/wxampunlink?access_token=TOKEN</a>
-   * 文档地址：
-   * <a href="https://developers.weixin.qq.com/doc/oplatform/Third-party_Platforms/2.0/api/Official__Accounts/Mini_Program_Management_Permission.html">https://developers.weixin.qq.com/doc/oplatform/Third-party_Platforms/2.0/api/Official__Accounts/Mini_Program_Management_Permission.html</a>
-   * <pre>
-   * @param appid 小程序 appid
-   * @return 响应结果
-   */
-  WxOpenResult wxAmpUnLink(String appid) throws WxErrorException;
 
   /**
    * 查询小程序版本信息
